requirements.txt
langchain
langchain_community
flask
python-box[all]~=7.0
PyYAML
langchain-chroma
pdfplumber
fastembed
langgraph
python-dotenv